updates docs and tests
