+++ README.md
@@ -2,6 +2,8 @@
 
 See [read the docs](https://gemini-medicine.github.io/pyDeid/index.html) for extensive examples and a detailed feature breakdown.
 
+**Note that `pyDeid` was tested with python `3.10.4`.**
+
 ## 1 Objective
 
 `PyDeid` is a Python-based refactor of the [Physionet Perl-based De-identification Software](https://physionet.org/content/deid/1.1/), which uses regular expressions and lookup dictionaries to identify and replace PHI in free text.
