+++ docs/api/index.rst
@@ -8,11 +8,3 @@ This section contains the auto-generated API documentation for pyDeid.
    :caption: API Contents:
 
    modules
-
-.. autosummary::
-   :toctree: _autosummary
-   :recursive:
-   :template: custom-module-template.rst
-   :caption: API Reference:
-
-   pyDeid
