+++ docs/tutorials/index.rst
@@ -8,6 +8,6 @@ This section contains tutorials for using pyDeid, presented as Jupyter notebooks
    :caption: Available Tutorials:
    :glob:
 
-  basic_usage.ipynb
-  using_spacy_with_pydeid.ipynb
-  evaluating_pydeid.ipynb
+   basic_usage.ipynb
+   using_spacy_with_pydeid.ipynb
+   evaluating_pydeid.ipynb
